Adding REPL  (#203)
This PR adds a copy of the Console REPL from Power-Fx that is Dataverse enabled.  

Ideally there would be a shared copy of the REPL somewhere between the two repos, but since this repo is not public at the moment, it is difficult to accomplish this.  IMHO, managing the single program.cs file isn't so bad and we can fix this later, it's more important to promote testing and debugging within the team now.

This PR adds two functions:
- DVConnect( string connection-string, bool multiorg-or-singleorg )
- DVAddTable( string symbol-table-name, string dataverse-logical-name-of-table )

Because DV tables are so very wide and field reference are very experience, an option has been added:
- Option( Options.FormatTableColumns, string column-names-comma-delimited )

For example:
```
> DVConnect( "<connection-string-to-dataverse-environment-with-sample-data>", true )
true

> DVAddTable( "Contacts", "contact" )
true

> Option( Options.FormatTableColumns, "address1_city,contactid,yomifullname" )
"address1_city,contactid,yomifullname"

> Set( x, FirstN( Contacts, 5 ) )
x:
  address1_city                                      contactid                 yomifullname
 =============== ============================================== ============================
          Miami   GUID("0e9130d4-3ef8-ed11-8847-6045bd081199")        Yvonne McKay (sample)
       Issaquah   GUID("109130d4-3ef8-ed11-8847-6045bd081199")   Susanna Stubberod (sample)
         Duvall   GUID("129130d4-3ef8-ed11-8847-6045bd081199")      Nancy Anderson (sample)
         Monroe   GUID("149130d4-3ef8-ed11-8847-6045bd081199")      Maria Campbell (sample)
       Issaquah   GUID("169130d4-3ef8-ed11-8847-6045bd081199")         Sidney Higa (sample)
```

Co-authored-by: anderson-joyle <joyle@outlook.com>
Co-authored-by: Luc Genetier <lucgen@microsoft.com>